Make ProducerIdManager an interface (#1061)
This is a preliminary refactor that makes ProducerIdManager an interface.
The plan is to contribute a new implementation of ProducerIdManager that does not use the MetadataStore service (ZooKeeper) but it stores the ids on a Pulsar topic.

Using Pulsar as backing store for generating the ids is better when you are using this feature in a MultiTenant environment and you want to fully isolate the data of each tenants.

(cherry picked from commit b07222d36f64ed64d722489cc3a1fb31fbafa420)
